Likely final update
This will probably be the most revised version of the code. A major bug was found and fixed. This was the x_val and y_val variables needing to be swapped due to the way in which C handles arrays. Instead of the initial "array[x-axis][y-axis]" I had thought it should have been, i had overlooked that it should instead be "array[y-axis][x-axis]". While the variable names could have been changed, it was faster to do this instead as i didn't have the time to go through the whole code this close to the deadline